Move `import_attr` util to `_common/` (#51316)
Quarantining library dependencies out of `_private/`

---------
