+++ python/ray/serve/config.py
@@ -14,7 +14,7 @@ from ray._private.pydantic_compat import (
     PrivateAttr,
     validator,
 )
-from ray._private.utils import import_attr
+from ray._common.utils import import_attr
 from ray.serve._private.constants import (
     DEFAULT_AUTOSCALING_POLICY,
     DEFAULT_GRPC_PORT,
