+++ python/ray/serve/scripts.py
@@ -14,7 +14,7 @@ import yaml
 
 import ray
 from ray import serve
-from ray._private.utils import import_attr
+from ray._common.utils import import_attr
 from ray.autoscaler._private.cli_logger import cli_logger
 from ray.dashboard.modules.dashboard_sdk import parse_runtime_env_args
 from ray.dashboard.modules.serve.sdk import ServeSubmissionClient
